refactor: domain 로직을 domain 폴더로 이동
